% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{alternating_squares_pointcloud_2D}
\alias{alternating_squares_pointcloud_2D}
\title{alternating_squares_pointcloud_2D}
\usage{
alternating_squares_pointcloud_2D(
  total_points = 1000,
  num_squares = 5,
  square_width = 0.5,
  spacing = 1,
  noise = 0.05
)
}
\arguments{
\item{total_points}{Total number of points to generate.}

\item{num_squares}{Number of squares.}

\item{square_width}{Width of each square.}

\item{spacing}{Spacing between squares.}

\item{noise}{Proportion of noise points to add.}
}
\value{
A numeric matrix of points (n x 2).
}
\description{
Generate test data: Alternating squares in 2D
}
