% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{alternating_squares_polygon}
\alias{alternating_squares_polygon}
\title{alternating_squares_polygon}
\usage{
alternating_squares_polygon(
  num_squares = 5,
  square_width = 0.5,
  spacing = 1,
  center_x = 0,
  center_y = 0
)
}
\arguments{
\item{num_squares}{Number of squares.}

\item{square_width}{Width of each square.}

\item{spacing}{Spacing between squares.}

\item{center_x}{X-coordinate of the center.}

\item{center_y}{Y-coordinate of the center.}
}
\value{
A list of matrices representing the polygons.
}
\description{
Generate polygon structure for alternating squares in 2D
}
