% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{area_concentric_rings_2D}
\alias{area_concentric_rings_2D}
\title{area_concentric_rings_2D}
\usage{
area_concentric_rings_2D(num_rings = 3, spacing = 1, ring_width = 0.5)
}
\arguments{
\item{num_rings}{Number of rings.}

\item{spacing}{Spacing between rings.}

\item{ring_width}{Width of each ring.}
}
\value{
The total area of the concentric rings.
}
\description{
Calculate area of concentric rings shape in 2D
}
