% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{auto_epsilon}
\alias{auto_epsilon}
\title{auto_epsilon}
\usage{
auto_epsilon(points, min_points = 5, quantile_value = 0.9)
}
\arguments{
\item{points}{A numeric matrix of points (n x d).}

\item{min_points}{The number of nearest neighbors to consider (k).}

\item{quantile_value}{The quantile of the k-NN distances to use as epsilon.}
}
\value{
A numeric value representing the estimated epsilon.
}
\description{
Estimate epsilon for DBSCAN using k-NN distance distribution
}
