% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{build_adj_my}
\alias{build_adj_my}
\title{build_adj_my}
\usage{
build_adj_my(edges)
}
\arguments{
\item{edges}{A list of numeric vectors length 2 (vertex indices).}
}
\value{
A named list where each name is a vertex and each element is a vector of adjacent vertices.
}
\description{
build adjacency list from edges
}
