% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{calculate_area_volume}
\alias{calculate_area_volume}
\title{calculate_area_volume}
\usage{
calculate_area_volume(alpha_shape, dim = 2)
}
\arguments{
\item{alpha_shape}{The output from \code{alpha_shape_2D} or \code{alpha_shape_3D}.}

\item{dim}{The dimension of the alpha shape (2 or 3).}
}
\value{
The area (2D) or volume (3D) of the alpha shape.
}
\description{
Calculate the area (2D) or volume (3D) of an alpha shape
}
