% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{calculate_surface_perimeter}
\alias{calculate_surface_perimeter}
\title{calculate_surface_perimeter}
\usage{
calculate_surface_perimeter(alpha_shape, dim = 2)
}
\arguments{
\item{alpha_shape}{The output from \code{alpha_shape_2D} or \code{alpha_shape_3D}.}

\item{dim}{The dimension of the alpha shape (2 or 3).}
}
\value{
The surface area (3D) or perimeter (2D) of the alpha shape.
}
\description{
Calculate surface area (3D) or perimeter (2D) of the alpha shape
}
