% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{circumcircle}
\alias{circumcircle}
\title{Circumcircle}
\usage{
circumcircle(p1, p2, p3)
}
\arguments{
\item{p1}{A vector of length 2 (x, y) coordinates of the first point.}

\item{p2}{A vector of length 2 (x, y) coordinates of the second point.}

\item{p3}{A vector of length 2 (x, y) coordinates of the third point.}
}
\value{
A list with elements 'center' (vector of length 2) and 'r' (radius).
}
\description{
Function to calculate the circumcircle of a triangle defined by points p1, p2, p3
}
