% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{circumsphere}
\alias{circumsphere}
\title{circumsphere}
\usage{
circumsphere(p1, p2, p3, p4)
}
\arguments{
\item{p1}{A vector of length 3 (x, y, z) coordinates of the first point.}

\item{p2}{A vector of length 3 (x, y, z) coordinates of the second point.}

\item{p3}{A vector of length 3 (x, y, z) coordinates of the third point.}

\item{p4}{A vector of length 3 (x, y, z) coordinates of the fourth point.}
}
\value{
A list with elements 'center' (vector of length 3) and 'r' (radius).
}
\description{
Function to calculate the circumsphere of a tetrahedron defined by points p1, p2, p3, p4
}
