% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{concentric_rings_2D}
\alias{concentric_rings_2D}
\title{concentric_rings_2D}
\usage{
concentric_rings_2D(
  num_rings = 3,
  spacing = 1,
  ring_width = 1,
  total_points = 1000,
  noise = 0.01
)
}
\arguments{
\item{num_rings}{Number of rings.}

\item{spacing}{Spacing between rings.}

\item{ring_width}{Width of each ring.}

\item{total_points}{Total number of points to generate.}

\item{noise}{Proportion of noise points to add.}
}
\value{
A numeric matrix of points (n x 2).
}
\description{
Generate test data: Concentric rings in 2D
}
