% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{corner_variations}
\alias{corner_variations}
\title{corner_variations}
\usage{
corner_variations(
  s = 3,
  w = 1.5,
  num_rings = 3,
  alpha_list = c(1),
  CL_list = c(0.95),
  num_points = 10000,
  noise_in_data_list = c(0.01)
)
}
\arguments{
\item{s}{Spacing between squares.}

\item{w}{Width of each square.}

\item{num_rings}{Number of rings or squares.}

\item{alpha_list}{List of alpha parameters for alpha shape.}

\item{CL_list}{List of confidence levels for alpha shape.}

\item{num_points}{Number of points to generate.}

\item{noise_in_data_list}{List of proportions of noise points to add.}
}
\value{
None. Saves results to CSV.
}
\description{
Run alpha shape tests varying corner shape parameters
}
