% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datapoints_variations}
\alias{datapoints_variations}
\title{datapoints_variations}
\usage{
datapoints_variations(num_points_list = c(10, 100, 1000, 10000, 1e+05, 1e+06))
}
\arguments{
\item{num_points_list}{List of number of points to test.}
}
\value{
None. Saves timing results to CSV.
}
\description{
Run alpha shape tests varying number of data points
}
