% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{dbscan_filter}
\alias{dbscan_filter}
\title{DBSCAN_filter}
\usage{
dbscan_filter(
  points,
  min_points = 5,
  epsilon = NULL,
  quantile_epsilon = 0.95,
  B = 2000,
  sample_fraction = 0.95,
  CL = 0.95
)
}
\arguments{
\item{points}{A numeric matrix of points (n x d).}

\item{min_points}{The minimum number of points to form a dense region.}

\item{epsilon}{The radius of the neighborhood to consider.}

\item{quantile_epsilon}{The quantile of the k-NN distances to use as epsilon if epsilon is NULL.}

\item{B}{The number of bootstrap samples.}

\item{sample_fraction}{The fraction of points to sample in each bootstrap.}

\item{CL}{The confidence level for filtering points.}
}
\value{
A matrix of filtered points.
}
\description{
DBSCAN-based filtering of points
}
