% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{edge_length}
\alias{edge_length}
\title{edge_length}
\usage{
edge_length(p1, p2)
}
\arguments{
\item{p1}{A numeric vector representing the first point (x, y).}

\item{p2}{A numeric vector representing the second point (x, y).}
}
\value{
The length of the edge.
}
\description{
Calculate the length of an edge given its two endpoints
}
