% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{edges_to_loops}
\alias{edges_to_loops}
\title{edges_to_loops}
\usage{
edges_to_loops(boundary_edges, points)
}
\arguments{
\item{boundary_edges}{A list of numeric vectors length 2 (vertex indices).}

\item{points}{A matrix of point coordinates (n x 2).}
}
\value{
A list of polygons, each polygon is a matrix (m x 2) of coordinates (closed: last == first).
}
\description{
Convert boundary edges to ordered loops (polygons)
}
