% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{error_in_2D}
\alias{error_in_2D}
\title{error_in_2D}
\usage{
error_in_2D(true_area, estimated_area)
}
\arguments{
\item{true_area}{The true area.}

\item{estimated_area}{The estimated area.}
}
\value{
The percentage error.
}
\description{
Calculate percentage error between true area and estimated area in 2D
}
