% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{find_cycles_my}
\alias{find_cycles_my}
\title{find_cycles_my}
\usage{
find_cycles_my(adj)
}
\arguments{
\item{adj}{A named list where each name is a vertex and each element is a
vector of adjacent vertices.}
}
\value{
A list of cycles, each cycle is a vector of vertex names.
}
\description{
find cycles in undirected graph
}
