% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{fix_orientation}
\alias{fix_orientation}
\title{fix_orientation}
\usage{
fix_orientation(mat, clockwise = FALSE)
}
\arguments{
\item{mat}{A matrix of polygon vertices (n x 2).}

\item{clockwise}{Logical indicating if the polygon should be oriented clockwise.}
}
\value{
A matrix of polygon vertices with the desired orientation.
}
\description{
Fix polygon orientation
}
