% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{heart_shape_datacloud}
\alias{heart_shape_datacloud}
\title{heart_shape_datacloud}
\usage{
heart_shape_datacloud(
  num_points = 1000,
  num_rings = 3,
  spacing = 1,
  ring_width = 0.5,
  noise = 0.05
)
}
\arguments{
\item{num_points}{Number of points to generate.}

\item{num_rings}{Number of rings.}

\item{spacing}{Spacing between rings.}

\item{ring_width}{Width of each ring.}

\item{noise}{Proportion of noise points to add.}
}
\value{
A numeric matrix of points (n x 2).
}
\description{
Generate test data: Heart shape in 2D
}
