% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pitchfork_bifurcation_datacloud}
\alias{pitchfork_bifurcation_datacloud}
\title{pitchfork_bifurcation_datacloud}
\usage{
pitchfork_bifurcation_datacloud(
  num_points = 1000,
  noise = 0.05,
  direction = "up"
)
}
\arguments{
\item{num_points}{Number of points to generate.}

\item{noise}{Proportion of noise points to add.}

\item{direction}{Direction of the bifurcation ("up" or "down").}
}
\value{
A numeric matrix of points (n x 2).
}
\description{
Generate test data: Pitchfork bifurcation in 2D
}
