% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{plot_datapoints_timing}
\alias{plot_datapoints_timing}
\title{plot_datapoints_timing}
\usage{
plot_datapoints_timing(
  csv_file = "alpha_shape_datapoints_timing.csv",
  scale = 1
)
}
\arguments{
\item{csv_file}{CSV file containing timing results.}

\item{scale}{Scale type for the plot (1: linear-linear, 2: linear-log, 3: log-log, 4: n vs sqrt t, 5: log n vs log sqrt t).}
}
\value{
None. Saves plot to PNG file.
}
\description{
Plot timing results for varying number of data points
}
