% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{point_in_circle}
\alias{point_in_circle}
\title{point_in_circle}
\usage{
point_in_circle(p, center, r)
}
\arguments{
\item{p}{A vector of length 2 (x, y) coordinates of the point.}

\item{center}{A vector of length 2 (x, y) coordinates of the circle center.}

\item{r}{Radius of the circle.}
}
\value{
TRUE if the point is inside the circle, FALSE otherwise.
}
\description{
Function to check if a point is inside a circle
}
