% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{point_in_sphere}
\alias{point_in_sphere}
\title{point_in_sphere}
\usage{
point_in_sphere(p, center, r)
}
\arguments{
\item{p}{A vector of length 3 (x, y, z) coordinates of the point.}

\item{center}{A vector of length 3 (x, y, z) coordinates of the sphere center.}

\item{r}{Radius of the sphere.}
}
\value{
TRUE if the point is inside the sphere, FALSE otherwise.
}
\description{
Function to check if a point is inside a sphere
}
