% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{polygon_area}
\alias{polygon_area}
\title{polygon_area}
\usage{
polygon_area(polygon)
}
\arguments{
\item{polygon}{A matrix of vertices (n x 2) representing the polygon.}
}
\value{
The area of the polygon.
}
\description{
Calculate the area of a polygon given its vertices
}
