% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{polygon_pitchfork}
\alias{polygon_pitchfork}
\title{polygon_pitchfork}
\usage{
polygon_pitchfork(x_min = -2, x_max = 2, r_min = -2, r_max = 2)
}
\arguments{
\item{x_min}{Minimum x-value.}

\item{x_max}{Maximum x-value.}

\item{r_min}{Minimum r-value.}

\item{r_max}{Maximum r-value.}
}
\value{
A matrix of polygon coordinates.
}
\description{
Generate polygon representing pitchfork bifurcation shape in 2D
}
