% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ring_test_generation_2D}
\alias{ring_test_generation_2D}
\title{ring_test_generation_2D}
\usage{
ring_test_generation_2D(
  num_points = 10000,
  r_outer = 3,
  r_inner = 1,
  noise = 0.05
)
}
\arguments{
\item{num_points}{Number of points to generate.}

\item{r_outer}{Outer radius.}

\item{r_inner}{Inner radius.}

\item{noise}{Proportion of noise points to add.}
}
\value{
A numeric matrix of points (n x 2).
}
\description{
Generate test data: Ring in 2D
}
