% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{square_area_2D}
\alias{square_area_2D}
\title{alternating_squares_area_2D}
\usage{
square_area_2D(
  num_squares = 5,
  square_width = 0.5,
  spacing = 1,
  center_x = 0,
  center_y = 0
)
}
\arguments{
\item{num_squares}{Number of squares.}

\item{square_width}{Width of each square.}

\item{spacing}{Spacing between squares.}

\item{center_x}{X-coordinate of the center.}

\item{center_y}{Y-coordinate of the center.}
}
\value{
The total area of the alternating squares.
}
\description{
Calculate area of alternating squares shape in 2D
}
