% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{tetrahedron_surface_area}
\alias{tetrahedron_surface_area}
\title{tetrahedron_surface_area}
\usage{
tetrahedron_surface_area(p1, p2, p3, p4)
}
\arguments{
\item{p1}{A numeric vector representing the first point (x, y, z).}

\item{p2}{A numeric vector representing the second point (x, y, z).}

\item{p3}{A numeric vector representing the third point (x, y, z).}

\item{p4}{A numeric vector representing the fourth point (x, y, z).}
}
\value{
The surface area of the tetrahedron.
}
\description{
Calculate the surface area of a tetrahedron given its four vertices
}
