% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{to_sf_poly}
\alias{to_sf_poly}
\title{to_sf_poly}
\usage{
to_sf_poly(rings)
}
\arguments{
\item{rings}{A list of matrices, each representing a ring (Mx2).}
}
\value{
An sf polygon object.
}
\description{
Convert list of rings to sf polygon
}
