% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{torus_test_generation_3D}
\alias{torus_test_generation_3D}
\title{torus_test_generation_3D}
\usage{
torus_test_generation_3D(num_points = 10000, R = 3, r = 1, noise = 0.05)
}
\arguments{
\item{num_points}{Number of points to generate.}

\item{R}{Major radius.}

\item{r}{Minor radius.}

\item{noise}{Standard deviation of Gaussian noise to add.}
}
\value{
A numeric matrix of points (n x 3).
}
\description{
Generate test data: Torus in 3D
}
