% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_shape.R
\name{triangle_area}
\alias{triangle_area}
\title{triangle_area}
\usage{
triangle_area(p1, p2, p3)
}
\arguments{
\item{p1}{A numeric vector representing the first point (x, y).}

\item{p2}{A numeric vector representing the second point (x, y).}

\item{p3}{A numeric vector representing the third point (x, y).}
}
\value{
The area of the triangle.
}
\description{
Calculate the area of a triangle given its three vertices
}
